# Reduce MAPPING_BASE_COST from 10_000 to 500