--- a/console/network/src/mainnet_v0.rs
+++ b/console/network/src/mainnet_v0.rs
@@ -136,7 +136,7 @@ impl Network for MainnetV0 {
 
     /// The block height from which consensus V2 rules apply.
     #[cfg(not(any(test, feature = "test")))]
-    const CONSENSUS_V2_HEIGHT: u32 = 2_000_000;
+    const CONSENSUS_V2_HEIGHT: u32 = 2_800_000;
     // TODO (raychu86): Update this value based on the desired mainnet height.
     /// The block height from which consensus V2 rules apply.
     #[cfg(any(test, feature = "test"))]