… ledger/query/src/query.rs
… ledger/query/src/traits.rs
… ledger/src/advance.rs
… ledger/src/check_next_block.rs
… synthesizer/Cargo.toml
… synthesizer/process/src/cost.rs
… synthesizer/process/src/stack/helpers/initialize.rs
… synthesizer/process/src/stack/mod.rs
… synthesizer/src/vm/execute.rs
… synthesizer/src/vm/finalize.rs
… synthesizer/src/vm/mod.rs
… synthesizer/src/vm/verify.rs
… synthesizer/tests/test_vm_execute_and_finalize.rs