# Lower GET, GET_OR_USE, & CONTAINS finalize fees

The goal of this PR is to lower fees safely such that the impact would encourage more chain activity without impacting the safety of network.
Specifically, the PR only changes a single value: `MAPPING_BASE_COST` from `10_000` microcredits to `500` microcredits

## Previous work

The most comphrensive work IMO is this analysis: https://github.com/iamalwaysuncomfortable/analysis/blob/main/aleo_finalize_opcode_benchmark_and_pricing_analysis.ipynb as part of this PR: https://github.com/AleoNet/snarkVM/pull/2281

Some key observations taken from this analysis:
 * SET & GET are the two most commonly used OPCodes
 * Most simple operations run in 1 to 5 µs and are priced at 500 microcredits
 * SET & GET have the same base cost of 10_000 credits with different per byte additional costs of 100 & 10 respectively.
 * The base assumption for the fee model is 1 second of single threaded runtime should cost 100 credits

## New Benchmarks Approach

I could not find the actual benchmarks for SET/GET in the linked previous work but all of the other opcodes are included.
So I created [these benchmarks for SET/GET](https://github.com/demox-labs/snarkVM/blob/c6bc485b199fde11124d8c1136d866e6559b5ec2/synthesizer/process/benches/update_key_value.rs) . I decided not to include these benchmarks as part of this PR but it can be used again when revisiting the fee model and as other DB optimizations land.

These new benchmarks use rocksdb & specifically benchmark: `finalize_store.update_key_value` & `finalize_store.get_value_speculative`.
Although this is the core part of a SET & GET operation, other overhead exists which is not included in this benchmark but the overhead should be roughly equivalent between SET & GET.

There are probably signficant differences from running benchmarks and running on a live node under heavy load but some care has been taken to ensure realistic benchmarks including using unique keys for the gets & sets and using rocksdb instead of the test finalize store.

## Results

Although priced similarly, GET & SET are extremely different in actual runtime.

For primitive types like a `U64` a SET costs `11_200` and a GET costs `10_120`. (A base cost of 10_000 and serialized, a U64 is 12 bytes)
The SET runtime for a `U64` is `170µs` and the GET runtime is `4.3µs`

### Benchmark Summary

- **Type: small_struct (422 bytes)**  
  - **Set Time:** ~440.30 µs  
  - **Get Time:** ~79.75 µs  

- **Type: large_struct (6456 bytes)**  
  - **Set Time:** ~4690 µs  
  - **Get Time:** ~135.55 µs  

- **Type: array_1x1 (43 bytes)**  
  - **Set Time:** ~185.21 µs  
  - **Get Time:** ~4.21 µs  

- **Type: array_32x32 (1407 bytes)**  
  - **Set Time:** ~1.23 ms  
  - **Get Time:** ~17.41 µs  

- **Type: U8**  
  - **Set Time:** ~165.74 µs  
  - **Get Time:** ~4.27 µs  

- **Type: U16**  
  - **Set Time:** ~169.60 µs  
  - **Get Time:** ~4.48 µs  

- **Type: U32**  
  - **Set Time:** ~154.07 µs  
  - **Get Time:** ~4.45 µs  

- **Type: U64**  
  - **Set Time:** ~168.84 µs  
  - **Get Time:** ~4.34 µs  

- **Type: U128**  
  - **Set Time:** ~170.26 µs  
  - **Get Time:** ~4.49 µs  



## Conclusions

GET is currently priced similarly to a SET when the in many cases, there is a 40x difference in runtime.
We should lower the base price of GET to `500` to put more inline with other primitive operations.

The big outlier is structs which perhaps deserve their own pricing model but I wanted to keep this change as simple as possible.

<img width="830" alt="Screenshot 2024-10-25 at 2 12 22 AM" src="https://github.com/user-attachments/assets/963146e8-dca4-4772-a858-0d57b87b70ea">

<img width="914" alt="Screenshot 2024-10-25 at 2 10 39 AM" src="https://github.com/user-attachments/assets/be548ec8-c518-449d-88cb-d5a049e5d2a1">


## Safety

Despite this change leading to a more consistent but still imperfect pricing model, we should consider the potential impact.

As a validator, we run a 32-core, 64GB memory computer to power our validator. Our CPU utilization for the node in the last 7 days has stayed between 5% and 10% despite relatively heavy load. It seem at current network activity levels, we are unlikely to risk overloading machines.

<img width="1575" alt="Screenshot 2024-10-23 at 6 23 57 PM" src="https://github.com/user-attachments/assets/27240c17-d807-46e5-815c-9b0c910f2779">


If there are other concerns, I'm unaware of please share.

## Impact

For many functions that use public state, we can expect to reduce the transaction fee by 35% to 60%. As the team behind the Leo Wallet, the most frequent compliant we hear is about the high fees.

To pick a couple of programs, I'm familiar with:
 - token_registry.aleo/transfer_public: 140_173 -> 92_673
 - pondo_protocol.aleo/deposit_public_as_signer: 477_484 -> 211_484

Since higher fees are still accepted, migration for downstream programs and applications isn't required, only highly recommended.

It will be difficult to raise fees in the future, so we should be aware that if we expect the runtime of GET to become more expensive over time, then we should perhaps be more cautious with how much we lower the GET base fees.