@vicsn 2024-10-25
As this would be the first PR which changes which transactions (and their fees) are deemed valid, we'll have to add some "migration" logic. Would defining it as follows suffice? The new pricing can apply from this height onwards. CC @raychu86 @d0cd 

```
diff --git a/console/network/src/canary_v0.rs b/console/network/src/canary_v0.rs
index f0a973acd..5c3770789 100644
--- a/console/network/src/canary_v0.rs
+++ b/console/network/src/canary_v0.rs
@@ -147,6 +147,8 @@ impl Network for CanaryV0 {
     const INCLUSION_FUNCTION_NAME: &'static str = MainnetV0::INCLUSION_FUNCTION_NAME;
     /// The maximum number of certificates in a batch.
     const MAX_CERTIFICATES: u16 = 100;
+    /// The block height from which new consensus rules apply.
+    const CONSENSUS_V2_HEIGHT: u32 = 1_000; // TODO: adjust based on canary height.
     /// The network name.
     const NAME: &'static str = "Aleo Canary (v0)";
 
diff --git a/console/network/src/lib.rs b/console/network/src/lib.rs
index b0694d43a..483157e15 100644
--- a/console/network/src/lib.rs
+++ b/console/network/src/lib.rs
@@ -203,6 +203,8 @@ pub trait Network:
 
     /// The maximum number of certificates in a batch.
     const MAX_CERTIFICATES: u16;
+    /// The block height from which new consensus rules apply.
+    const CONSENSUS_V2_HEIGHT: u32;
 
     /// The maximum number of bytes in a transaction.
     // Note: This value must **not** be decreased as it would invalidate existing transactions.
diff --git a/console/network/src/mainnet_v0.rs b/console/network/src/mainnet_v0.rs
index 6aadaa3f3..1aaaecc9e 100644
--- a/console/network/src/mainnet_v0.rs
+++ b/console/network/src/mainnet_v0.rs
@@ -158,6 +158,8 @@ impl Network for MainnetV0 {
     const INCLUSION_FUNCTION_NAME: &'static str = snarkvm_parameters::mainnet::NETWORK_INCLUSION_FUNCTION_NAME;
     /// The maximum number of certificates in a batch.
     const MAX_CERTIFICATES: u16 = 16;
+    /// The block height from which new consensus rules apply.
+    const CONSENSUS_V2_HEIGHT: u32 = 3_000_000; // TODO: adjust based on mainnet height.
     /// The network name.
     const NAME: &'static str = "Aleo Mainnet (v0)";
 
diff --git a/console/network/src/testnet_v0.rs b/console/network/src/testnet_v0.rs
index fd793055a..0f42c7c90 100644
--- a/console/network/src/testnet_v0.rs
+++ b/console/network/src/testnet_v0.rs
@@ -147,6 +147,8 @@ impl Network for TestnetV0 {
     const INCLUSION_FUNCTION_NAME: &'static str = MainnetV0::INCLUSION_FUNCTION_NAME;
     /// The maximum number of certificates in a batch.
     const MAX_CERTIFICATES: u16 = 100;
+    /// The block height from which new consensus rules apply.
+    const CONSENSUS_V2_HEIGHT: u32 = 1_000; // TODO:  adjust based on testnet height.
     /// The network name.
     const NAME: &'static str = "Aleo Testnet (v0)";
 ```