@evanmarshall 2024-10-25
> As this would be the first PR which changes which transactions (and their fees) are deemed valid

Do we have to add a migration? All previously valid transactions will still be valid.