@vicsn 2024-10-25
> > As this would be the first PR which changes which transactions (and their fees) are deemed valid

> Do we have to add a migration? All previously valid transactions will still be valid.

Good point. I still think it would be good for:
- Defense in depth. I'm wondering whether this could make`rejected` transactions could potentially be considered valid (don't know the potential reasons for rejection by heart), and I want to make reviewers life easy by not having to think about this.
- Avoiding a synchronous restart. If only some validators have the upgrade, it could lead to a fork.