@raychu86 2024-10-25
> Good point. I still think it would be good for:
> 
> * Defense in depth. I'm wondering whether this could make`rejected` transactions could potentially be considered valid (don't know the potential reasons for rejection by heart), and I want to make reviewers life easy by not having to think about this.
> * Avoiding a synchronous restart. If only some validators have the upgrade, it could lead to a fork.

1. _Technically_ there is no requirement to add additional migration logic for pricing reductions because previously, txes that have too few fees either never make it to DAG serialization or are aborted by honest validators. (Note that is not the case for pricing increases, because it could invalidate accepted txes.). **However**, because mainnet is live and we can not expect a coordinated restart of all nodes, we need to consider adding logic s.t. nodes automatically roll into the new change once it passes a future block X that we specify. 

2. Any consensus change (which this is) should increment the snarkOS message versions, so nodes on older versions do not  connect to nodes on the new version. This is our defense against forking, however it's a bit more nuanced.

TLDR: If we do a coordinated restart, then we can do without extra migration logic. Otherwise, migration logic may be necessary to keep the network rolling