@evanmarshall 2024-10-25
> TLDR: If we do a coordinated restart, then we can do without extra migration logic. Otherwise, migration logic may be necessary to keep the network rolling

In any case, we should plan on a synchronous upgrade as well to minimize any chance of a chain halt because during an async upgrade, it's likely neither version will have quorum. If we plan on a synchronous upgrade, do you think the migration is necessary? I'm happy to add it but I'm not sure if it's preferred to keep that branching code around forever.

> In general, since I assume there will be a larger effort to reduce fees overall other parts of finalize/storage they should be implemented together to avoid multiple migration efforts. Additionally, I would advise a fee reduction discussion take place as an ARC so there is more weigh in regarding pain points (and to discuss the migration considerations as well).

I absolutely think we should examine every part of fees in general but I strongly suggest we implement this change immediately while we consider other larger changes. I don't think it requires an ARC as none of the mechanisms, logic, or assumptions are changing. The only part that's changing is fixing the inconsistency of pricing `GET/GET_OR_USE/CONTAINS`.