@vicsn 2024-10-28
I agree with Evan that given this change does not change assumptions and is urgent, it is unlikely to warrant an ARC. Future bigger fee adjustment proposals likely will require broader discussion. But its up to Foundation to draw the line. CC @zkxuerb 

I agree with Ray there should be gateway `EVENT` version bump. Want to add this in @evanmarshall ? Also removes the *need* for migration logic.

A router `MESSAGE` version bump would force a coordinated restart for all the clients in the ecosystem, which sounds very difficult to coordinate. Other approaches would be:
1. Use migration logic, which gives clients a larger window to update.
2. Don't use migration logic nor do a `MESSAGE` version bump. Increasing the risk that old clients will temporarily relay transactions with incorrect fees (which is unfortunate but acceptable) or risk forking because of syncing with incorrect blocks (I did not analyse in depth yet).