@d0cd 2024-10-31
Do we have data (either in this PR or prior) that helps us understand:
1. How the performance of `get` varies as the size of the data type increases? 
2. For gets of data type of varying sizes, how does the performance vary as the number of entries in the mapping increases?
I believe the current benchmarks are on small-sized gets, which runs us the risk of missing a DOS vector at larger sizes.